backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.6
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1